+++ components/collector/src/source_collectors/junit/source_up_to_dateness.py
@@ -2,8 +2,10 @@
 
 from datetime import datetime
 
+from shared.utils.date_time import now
+
 from base_collectors import TimePassedCollector, XMLFileSourceCollector
-from collector_utilities.date_time import now, parse_datetime
+from collector_utilities.date_time import parse_datetime
 from collector_utilities.functions import parse_source_response_xml
 from collector_utilities.type import Response