+++ components/collector/src/source_collectors/quality_time/metrics.py
@@ -4,9 +4,10 @@ from datetime import datetime
 from typing import cast
 from urllib import parse
 
+from shared.utils.date_time import now
 from shared_data_model import DATA_MODEL
 
-from collector_utilities.date_time import now, parse_datetime
+from collector_utilities.date_time import parse_datetime
 from collector_utilities.type import URL, Response, Value
 from model import Entities, Entity, SourceMeasurement, SourceResponses