+++ components/collector/tests/collector_utilities/test_date_time.py
@@ -5,26 +5,15 @@ from datetime import datetime, timedelta
 
 from dateutil.tz import tzlocal, tzutc
 
+from shared.utils.date_time import now
+
 from collector_utilities.date_time import (
     datetime_fromparts,
     datetime_fromtimestamp,
     days_ago,
     days_to_go,
-    now,
     parse_datetime,
 )
-
-
-class NowTest(unittest.TestCase):
-    """Unit tests for the now function."""
-
-    def test_tz(self):
-        """Test that the timezone of now is equal to the local timezone."""
-        self.assertEqual(tzlocal(), now().tzinfo)
-
-    def test_now_equals_datetime_now(self):
-        """Test that now is equal to datetime.now() with the local timezone."""
-        self.assertTrue(now() - datetime.now(tz=tzlocal()) < timedelta(seconds=1))
 
 
 class DaysAgoTest(unittest.TestCase):