+++ components/frontend/src/defaults.js
@@ -1,7 +1,11 @@
 // Note that the API-server also has these constants, in src/model/report.py.
-export const metricReactionDeadline = {
+export const defaultDesiredResponseTimes = {
     debt_target_met: 60,
     near_target_met: 21,
     target_not_met: 7,
     unknown: 3,
+    confirmed: 180,
+    false_positive: 180,
+    wont_fix: 180,
+    fixed: 7,
 }