+++ components/frontend/src/report/ReportTitle.test.js
@@ -65,7 +65,7 @@ it('sets the unknown status reaction time', async () => {
     render_report_title();
     fireEvent.click(screen.getByTitle(/expand/));
     await act(async () => { fireEvent.click(screen.getByText(/reaction times/)) });
-    await userEvent.type(screen.getByLabelText(/unknown status/), '4{Enter}}', { initialSelectionStart: 0, initialSelectionEnd: 1 });
+    await userEvent.type(screen.getByLabelText(/Unknown/), '4{Enter}}', { initialSelectionStart: 0, initialSelectionEnd: 1 });
     expect(report_api.set_report_attribute).toHaveBeenLastCalledWith("report_uuid", "desired_response_times", { "unknown": 4 }, reload);
 })
 
@@ -73,7 +73,7 @@ it('sets the target not met status reaction time', async () => {
     render_report_title();
     fireEvent.click(screen.getByTitle(/expand/));
     await act(async () => { fireEvent.click(screen.getByText(/reaction times/)) });
-    await userEvent.type(screen.getByLabelText(/not meeting their target/), '5{Enter}}', { initialSelectionStart: 0, initialSelectionEnd: 1 });
+    await userEvent.type(screen.getByLabelText(/Target not met/), '5{Enter}}', { initialSelectionStart: 0, initialSelectionEnd: 1 });
     expect(report_api.set_report_attribute).toHaveBeenLastCalledWith("report_uuid", "desired_response_times", { "target_not_met": 5 }, reload);
 })
 
@@ -81,7 +81,7 @@ it('sets the near target met status reaction time', async () => {
     render_report_title();
     fireEvent.click(screen.getByTitle(/expand/));
     await act(async () => { fireEvent.click(screen.getByText(/reaction times/)) });
-    await userEvent.type(screen.getByLabelText(/near their target/), '6{Enter}}', { initialSelectionStart: 0, initialSelectionEnd: 2 });
+    await userEvent.type(screen.getByLabelText(/Near target met/), '6{Enter}}', { initialSelectionStart: 0, initialSelectionEnd: 2 });
     expect(report_api.set_report_attribute).toHaveBeenLastCalledWith("report_uuid", "desired_response_times", { "near_target_met": 6 }, reload);
 })
 
@@ -89,8 +89,40 @@ it('sets the tech debt target status reaction time', async () => {
     render_report_title();
     fireEvent.click(screen.getByTitle(/expand/));
     await act(async () => { fireEvent.click(screen.getByText(/reaction times/)) });
-    await userEvent.type(screen.getByLabelText(/accepted technical debt/), '6{Enter}}', { initialSelectionStart: 0, initialSelectionEnd: 2 });
+    await userEvent.type(screen.getByLabelText(/Technical debt target met/), '6{Enter}}', { initialSelectionStart: 0, initialSelectionEnd: 2 });
     expect(report_api.set_report_attribute).toHaveBeenLastCalledWith("report_uuid", "desired_response_times", { "debt_target_met": 6 }, reload);
+})
+
+it('sets the confirmed measurement entity status reaction time', async () => {
+    render_report_title();
+    fireEvent.click(screen.getByTitle(/expand/));
+    await act(async () => { fireEvent.click(screen.getByText(/reaction times/)) });
+    await userEvent.type(screen.getByLabelText(/Confirmed/), '60{Enter}}', { initialSelectionStart: 0, initialSelectionEnd: 3 });
+    expect(report_api.set_report_attribute).toHaveBeenLastCalledWith("report_uuid", "desired_response_times", { "confirmed": 60 }, reload);
+})
+
+it('sets the false positive measurement entity status reaction time', async () => {
+    render_report_title();
+    fireEvent.click(screen.getByTitle(/expand/));
+    await act(async () => { fireEvent.click(screen.getByText(/reaction times/)) });
+    await userEvent.type(screen.getByLabelText(/False positive/), '70{Enter}}', { initialSelectionStart: 0, initialSelectionEnd: 3 });
+    expect(report_api.set_report_attribute).toHaveBeenLastCalledWith("report_uuid", "desired_response_times", { "false_positive": 70 }, reload);
+})
+
+it('sets the will be fixed measurement entity status reaction time', async () => {
+    render_report_title();
+    fireEvent.click(screen.getByTitle(/expand/));
+    await act(async () => { fireEvent.click(screen.getByText(/reaction times/)) });
+    await userEvent.type(screen.getByLabelText(/Will be fixed/), '80{Enter}}', { initialSelectionStart: 0, initialSelectionEnd: 3 });
+    expect(report_api.set_report_attribute).toHaveBeenLastCalledWith("report_uuid", "desired_response_times", { "fixed": 80 }, reload);
+})
+
+it("sets the won't fixed measurement entity status reaction time", async () => {
+    render_report_title();
+    fireEvent.click(screen.getByTitle(/expand/));
+    await act(async () => { fireEvent.click(screen.getByText(/reaction times/)) });
+    await userEvent.type(screen.getByLabelText(/Won't fix/), '90{Enter}}', { initialSelectionStart: 0, initialSelectionEnd: 3 });
+    expect(report_api.set_report_attribute).toHaveBeenLastCalledWith("report_uuid", "desired_response_times", { "wont_fix": 90 }, reload);
 })
 
 it('sets the issue tracker type', async () => {