+++ components/frontend/src/source/SourceEntityDetails.js
@@ -15,17 +15,23 @@ function entity_status_option(status, text, content, subheader) {
     }
 }
 
-function entity_status_options(entity_type) {
+function entity_status_options(entity_type, report) {
+    const desired_response_times = report?.desired_response_times ?? {}
+    const unconfirmed_subheader = `This ${entity_type} should be reviewed to decide what to do with it.`
+    const confirmed_subheader = `This ${entity_type} has been reviewed and should be dealt with within ${desired_response_times["confirmed"]} days.`
+    const fixed_subheader = `Ignore this ${entity_type} for ${desired_response_times["fixed"]} days because it will be fixed shortly.`
+    const false_positive_subheader = `Ignore this ${entity_type} for ${desired_response_times["false_positive"]} days because it's been incorrectly identified as ${entity_type}.`
+    const wont_fix_subheader = `Ignore this ${entity_type} for ${desired_response_times["wont_fix"]} days because it will not be fixed.`
     return [
-        entity_status_option('unconfirmed', status_name.unconfirmed, 'Unconfirm', `This ${entity_type} should be reviewed to decide what to do with it.`),
-        entity_status_option('fixed', status_name.fixed, "Resolve as will be fixed", `This ${entity_type} will be fixed shortly and then disappear.`),
-        entity_status_option('false_positive', status_name.false_positive, 'Resolve as false positive', `This ${entity_type} can be ignored because it's been incorrectly identified as ${entity_type}.`),
-        entity_status_option('wont_fix', status_name.wont_fix, "Resolve as won't fix", `This ${entity_type} will not be fixed.`),
-        entity_status_option('confirmed', status_name.confirmed, 'Confirm', `This ${entity_type} has been reviewed and should be dealt with.`),
+        entity_status_option('unconfirmed', status_name.unconfirmed, 'Unconfirm', unconfirmed_subheader),
+        entity_status_option('confirmed', status_name.confirmed, "Confirm", confirmed_subheader),
+        entity_status_option('fixed', status_name.fixed, "Resolve as will be fixed", fixed_subheader),
+        entity_status_option('false_positive', status_name.false_positive, "Resolve as false positive", false_positive_subheader),
+        entity_status_option('wont_fix', status_name.wont_fix, "Resolve as won't fix", wont_fix_subheader),
     ]
 }
 
-export function SourceEntityDetails({ entity, metric_uuid, name, rationale, reload, status, status_end_date, source_uuid }) {
+export function SourceEntityDetails({ entity, metric_uuid, name, rationale, reload, report, status, status_end_date, source_uuid }) {
     return (
         <Grid stackable>
             <Grid.Row>
@@ -33,7 +39,7 @@ export function SourceEntityDetails({ entity, metric_uuid, name, rationale, relo
                     <SingleChoiceInput
                         requiredPermissions={[EDIT_ENTITY_PERMISSION]}
                         label={`${capitalize(name)} status`}
-                        options={entity_status_options(name)}
+                        options={entity_status_options(name, report)}
                         set_value={(value) => set_source_entity_attribute(metric_uuid, source_uuid, entity.key, "status", value, reload)}
                         value={status}
                         sort={false}