+++ components/frontend/src/utils.js
@@ -1,7 +1,7 @@
 import { useEffect, useState } from 'react';
 import { PERMISSIONS } from './context/Permissions';
 import { HyperLink } from './widgets/HyperLink';
-import { metricReactionDeadline } from './defaults';
+import { defaultDesiredResponseTimes } from './defaults';
 
 export const MILLISECONDS_PER_HOUR = 60 * 60 * 1000;
 const MILLISECONDS_PER_DAY = 24 * MILLISECONDS_PER_HOUR;
@@ -72,13 +72,14 @@ export function getMetricUnit(metric, dataModel) {
 
 export function getMetricResponseDeadline(metric, report) {
     let deadline = null;
-    if (metric.status === "debt_target_met") {
+    const status = metric.status || "unknown"
+    if (status === "debt_target_met") {
         if (metric.debt_end_date) {
             deadline = new Date(metric.debt_end_date)
         }
-    } else if ((metric.status || "unknown") in metricReactionDeadline && metric.status_start) {
+    } else if (status in defaultDesiredResponseTimes && metric.status_start) {
         deadline = new Date(metric.status_start)
-        deadline.setDate(deadline.getDate() + getMetricDesiredResponseTime(report, metric.status))
+        deadline.setDate(deadline.getDate() + getMetricDesiredResponseTime(report, status))
     }
     return deadline
 }
@@ -113,7 +114,7 @@ export function getMetricResponseOverrun(metric_uuid, metric, report, measuremen
     const overruns = []
     consolidatedMeasurements.forEach((measurement) => {
         const status = measurement?.[scale]?.status || "unknown"
-        if (status in metricReactionDeadline) {
+        if (status in defaultDesiredResponseTimes) {
             const desiredResponseTime = getMetricDesiredResponseTime(report, status) * MILLISECONDS_PER_DAY;
             const actualResponseTime = (new Date(measurement.end)).getTime() - (new Date(measurement.start)).getTime()
             const overrun = Math.max(0, actualResponseTime - desiredResponseTime)
@@ -136,7 +137,8 @@ export function getMetricResponseOverrun(metric_uuid, metric, report, measuremen
 }
 
 function getMetricDesiredResponseTime(report, status) {
-    return report?.desired_response_times?.[status] ?? (metricReactionDeadline[status] ?? metricReactionDeadline["unknown"])
+    // Precondition: status is a key of defaultDesiredResponseTimes
+    return report?.desired_response_times?.[status] ?? defaultDesiredResponseTimes[status]
 }
 
 export function get_metric_value(metric) {